refactor: move ProgressIndicator into components
