+++ frontend/src/components/ProgressIndicator/ProgressIndicator.tsx
